+++ raiden/utils/packing.py
